Update README.md
I upload two projects here. The first one is a house price prediction system. A lot of different factors are considered here to make it able to predict if the house will be expensive or not. The second one is one extra test data from another quite similar test to check the accuracy even to some small changes in variables. In both projects are the decision tree method applied.